requirements.txt
opencv-python
Pillow
natsort
numpy
supervision
pandas
matplotlib
seaborn
rfdetr